moving functions from <>AnimalSelectionDisplays to AnimalRouletteDisplay
